+++ cmd/carrel/main.go
@@ -29,7 +29,7 @@ import (
 )
 
 var (
-	version = "0.8.0"
+	version = "0.9.0"
 	commit  = "unknown"
 )